Rename plugin into amd_rocm_smi
