+++ etc/telegraf.conf
@@ -5786,7 +5786,7 @@
 #   servers = ["http://localhost:8098"]
 
 # # Query statistics from AMD Graphics cards using rocm-smi binary
-# [[inputs.rocm_smi]]
+# [[inputs.amd_rocm_smi]]
 # ## Optional: path to rocm-smi binary, defaults to $PATH via exec.LookPath
 # # bin_path = "/opt/rocm/bin/rocm-smi"
 #
