+++ plugins/inputs/amd_rocm_smi/testdata/vega-10-XT.json
