+++ plugins/inputs/amd_rocm_smi/testdata/vega-20-WKS-GL-XE.json
